switched to doubles to decrease dsync

+++ src/audio/Beat.java
@@ -5,23 +5,23 @@ import java.util.LinkedList;
 public class Beat {
 
    private float dst = 0.0f;
-   private LinkedList<Float> beats = null;
+   private LinkedList<Double> beats = null;
    private int currBeat = 0;
 
-   public Beat(LinkedList<Float> beats) {
+   public Beat(LinkedList<Double> beats) {
       this.beats = beats;
    }
 
-   public float peek() {
+   public double peek() {
       return this.beats.get(currBeat);
    }
 
-   public float time() {
+   public double time() {
       if (currBeat == 0) return beats.get(currBeat);
       return beats.get(currBeat) - beats.get(currBeat - 1);
    }
 
-   public LinkedList<Float> getList() {
+   public LinkedList<Double> getList() {
       return this.beats;
    }
